app/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.1
Appium-Python-Client==4.5.0
attrs==24.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
config==0.5.1
Deprecated==1.2.18
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
frozenlist==1.5.0
greenlet==3.2.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
propcache==0.3.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
rich==13.9.4
schedule==1.2.2
selenium==4.28.0
setuptools==79.0.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
trio==0.28.0
trio-websocket==0.11.1
twilio==9.4.6
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
